+++ src/api/user_data.ts
@@ -5,19 +5,22 @@
 import { mastodon } from "masto";
 
 import Account from "./objects/account";
+// import BooleanFilterOptionList from "../filters/boolean_filter_option_list";
 import MastoApi from "./api";
-import MostFavouritedAccountsScorer from "../scorer/feature/most_favourited_accounts_scorer";
-import MostRetootedAccountsScorer from "../scorer/feature/most_retooted_accounts_scorer";
-import Storage from "../Storage";
+// import MostFavouritedAccountsScorer from "../scorer/feature/most_favourited_accounts_scorer";
+// import MostRetootedAccountsScorer from "../scorer/feature/most_retooted_accounts_scorer";
 import ObjWithCountList, { ObjList } from "./obj_with_counts_list";
+import Storage from "../Storage";
 import TagList from "./tag_list";
 import Toot from "./objects/toot";
+import { BooleanFilterName } from '../enums';
 import { CacheKey, ScoreName, TagTootsCacheKey } from "../enums";
 import { config } from "../config";
-import { addDicts, countValues, sortKeysByValue } from "../helpers/collection_helpers";
+import { addDicts, countValues, keyById, sortKeysByValue } from "../helpers/collection_helpers";
 import { Logger } from '../helpers/logger';
 import {
     type AccountNames,
+    type BooleanFilterOption,
     type StringNumberDict,
     type TagWithUsageCounts
 } from "../types";
@@ -36,43 +39,68 @@ interface UserApiData {
 
 
 export default class UserData {
-    favouriteAccounts: ObjList = new ObjWithCountList([], ScoreName.FAVOURITED_ACCOUNTS);
-    favouritedTags: TagList = new TagList([], TagTootsCacheKey.FAVOURITED_TAG_TOOTS);
-    followedAccounts: StringNumberDict = {};  // Don't store the Account objects, just webfingerURI to save memory
-    followedTags: TagList = new TagList([], ScoreName.FOLLOWED_TAGS);
-    languagesPostedIn: StringNumberDict = {};
+    // numToots in favouriteAccounts is the sum of retoots, favourites, and replies to that account
+    favouriteAccounts = new ObjWithCountList<BooleanFilterOption>([], ScoreName.FAVOURITED_ACCOUNTS);
+    favouritedTags = new TagList([], TagTootsCacheKey.FAVOURITED_TAG_TOOTS);
+    followedAccounts: StringNumberDict = {};
+    followedTags = new TagList([], ScoreName.FOLLOWED_TAGS);
+    languagesPostedIn: ObjList = new ObjWithCountList([], BooleanFilterName.LANGUAGE);
     mutedAccounts: AccountNames = {};
-    participatedTags: TagList = new TagList([], TagTootsCacheKey.PARTICIPATED_TAG_TOOTS);
-    preferredLanguage: string = config.locale.defaultLanguage;
+    participatedTags = new TagList([], TagTootsCacheKey.PARTICIPATED_TAG_TOOTS);
+    preferredLanguage = config.locale.defaultLanguage;
     serverSideFilters: mastodon.v2.Filter[] = [];  // TODO: currently unused, only here for getCurrentState() by client app
 
     // Alternate constructor to build UserData from raw API data
     static buildFromData(data: UserApiData): UserData {
         const userData = new UserData();
+        userData.favouriteAccounts = this.buildFavouriteAccount(data);
         userData.favouritedTags = TagList.fromUsageCounts(data.favouritedToots, TagTootsCacheKey.FAVOURITED_TAG_TOOTS);
         userData.followedAccounts = Account.countAccounts(data.followedAccounts);
         userData.followedTags = new TagList(data.followedTags, ScoreName.FOLLOWED_TAGS);
-        userData.languagesPostedIn = countValues<Toot>(data.recentToots, (toot) => toot.language);
         userData.mutedAccounts = Account.buildAccountNames(data.mutedAccounts);
         userData.participatedTags = TagList.fromUsageCounts(data.recentToots, TagTootsCacheKey.PARTICIPATED_TAG_TOOTS);
-        userData.preferredLanguage = sortKeysByValue(userData.languagesPostedIn)[0] || config.locale.defaultLanguage;
         userData.serverSideFilters = data.serverSideFilters;
 
-        // Add up the favourites, retoots, and replies for each account
-        // TODO: can't include replies yet bc we don't have the webfingerURI for those accounts, only inReplyToID
-        const favouritedAccounts = MostFavouritedAccountsScorer.buildFavouritedAccounts(data.favouritedToots);
-        const retootedAccounts = MostRetootedAccountsScorer.buildRetootedAccounts(data.recentToots);
-
-        // Fill in zeros for accounts that the user follows but has not favourited or retooted
-        const followedAccountZeros = data.followedAccounts.reduce((zeros, account) => {
-            zeros[account.webfingerURI] = 0;
-            return zeros;
-        }, {} as StringNumberDict);
-
-        const accountsDict = addDicts(favouritedAccounts, followedAccountZeros, retootedAccounts);
-        userData.favouriteAccounts = ObjWithCountList.buildFromDict(accountsDict, ScoreName.FAVOURITED_ACCOUNTS);
+        // Language stuff
+        const languageUsageCounts = countValues(data.recentToots, (toot) => toot.language);
+        userData.languagesPostedIn = ObjWithCountList.buildFromDict(languageUsageCounts, BooleanFilterName.LANGUAGE);
+        userData.preferredLanguage = userData.languagesPostedIn.topObjs()[0]?.name || config.locale.defaultLanguage;
         logger.trace("Built from data:", userData);
         return userData;
+    }
+
+    // Add up the favourites, retoots, and replies for each account
+    private static buildFavouriteAccount(data: UserApiData): ObjWithCountList<BooleanFilterOption> {
+        const retootsAndFaves = [...Toot.onlyRetoots(data.recentToots), ...data.favouritedToots];
+        const retootAndFaveAccounts = retootsAndFaves.map(t => t.account);
+        const followedAccountIdMap = keyById(data.followedAccounts);
+
+        // TODO: Replies are imperfect - we're only checking followed accts bc we only have account ID to work with
+        const repliedToAccounts = Toot.onlyReplies(data.recentToots)
+                                      .map(toot => followedAccountIdMap[toot.inReplyToAccountId!])
+                                      .filter(Boolean);
+
+        const accountCounts = Account.countAccountsWithObj([...repliedToAccounts, ...retootAndFaveAccounts]);
+
+        // Fill in zeros for accounts that the user follows but has not favourited or retooted
+        data.followedAccounts.forEach((account) => {
+            accountCounts[account.webfingerURI] ??= {account, count: 0};
+            accountCounts[account.webfingerURI].isFollowed = true;
+        });
+
+        const accountOptions = Object.values(accountCounts).map(accountCount => {
+            const option: BooleanFilterOption = {
+                displayName: accountCount.account.displayName,
+                displayNameWithEmoji: accountCount.account.displayNameWithEmojis(),
+                isFollowed: accountCount.isFollowed,
+                name: accountCount.account.webfingerURI,
+                numToots: accountCount.count,
+            }
+
+            return option;
+        });
+
+        return new ObjWithCountList(accountOptions, ScoreName.FAVOURITED_ACCOUNTS);
     }
 
     // Alternate constructor for the UserData object to build itself from the API (or cache)