+++ src/enums.ts
@@ -85,8 +85,35 @@ export enum TrendingType {
     TAGS = "tags"
 };
 
-// Both filter option property names as well as demo app gradient config keys
-export const FILTER_OPTION_DATA_SOURCES = [
-    ...Object.values(TagTootsCacheKey),  // TODO: these are really the wrong cache keys for the use case but it's consistent w/demo app for now
-    ScoreName.FAVOURITED_ACCOUNTS,
-] as const;
+
+// Filters
+export enum BooleanFilterName {
+    HASHTAG = 'hashtag',
+    LANGUAGE = 'language',
+    TYPE = 'type',
+    USER = 'user',
+    APP = 'app'
+};
+
+// The values have spaces to make them more usable in the demo app's presentation
+export enum TypeFilterName {
+    AUDIO = 'audio',
+    BOT = 'bot',
+    DIRECT_MESSAGE = 'direct messages',
+    FOLLOWED_ACCOUNTS = 'followed accounts',
+    FOLLOWED_HASHTAGS = 'followed hashtags',
+    IMAGES = 'images',
+    LINKS = 'links',
+    MENTIONS = 'mentions',
+    PARTICIPATED_TAGS = 'participated hashtags',
+    POLLS = 'polls',
+    PRIVATE = 'private',
+    REPLIES = 'replies',
+    RETOOTS = 'retoots',
+    SENSITIVE = 'sensitive',
+    SPOILERED = 'spoilered',
+    TRENDING_LINKS = 'trending links',
+    TRENDING_TAGS = 'trending hashtags',
+    TRENDING_TOOTS = 'trending toots',
+    VIDEOS = 'videos'
+};